[ASTS] Targeted Sweeper Part 4: Separate out the sweeper
